Fix MSVC compile errors and correctness warnings
First real cl.exe build surfaced GCC-vs-MSVC issues:

Errors:
  - alea_bitset.h used __builtin_popcountll (GCC-only). Added a portable
    alea_popcount64(): SWAR software popcount under MSVC, __builtin_popcountll
    elsewhere. SWAR (not __popcnt64) avoids depending on the POPCNT
    instruction, matching GCC's portable default.
  - mcnp_conversion.c: `static const double FAR = 9.9e5;` is not a constant
    expression in standard C, so MSVC rejects it as a static-array initializer
    (C2099; GCC allows it as an extension). Made it a macro, renamed to
    FAR_DISTANCE to dodge the legacy FAR macro from <windows.h>.

Correctness (MSVC is LLP64: long is 32-bit):
  - geom_parser.c / geom_lexer.c printed a 64-bit ptrdiff_t with %ld (C4477),
    which misformats on MSVC. Use %lld with a (long long) cast.

Warning cleanup:
  - alea_spatial.c: dropped a top-level `restrict` on query_callback's
    parameter so its type matches the alea_spatial_callback typedef exactly
    (MSVC C4113; GCC ignores the qualifier for pointer compatibility).

GCC build unchanged: 683/0 tests pass.

+++ src/mcnp/parser/geom_lexer.c
@@ -237,7 +237,7 @@ const geometry_token_t* geom_lexer_next_token(geometry_lexer_t* lexer) {
             // Parsing failed
             lexer->token.type = GEOM_TOKEN_ERROR;
             snprintf(lexer->error_msg, sizeof(lexer->error_msg), 
-                    "Invalid number at position %ld", number_start - lexer->input);
+                    "Invalid number at position %lld", (long long)(number_start - lexer->input));
             lexer->token.end = lexer->current;
             return &lexer->token;
         }
